+++ projects/tinymce/src/lib/template-editor/template-editor-inner/template-editor-inner.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { TemplateEditorInnerComponent } from './template-editor-inner.component';
+
+describe('TemplateEditorInnerComponent', () => {
+  let component: TemplateEditorInnerComponent;
+  let fixture: ComponentFixture<TemplateEditorInnerComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ TemplateEditorInnerComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(TemplateEditorInnerComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});