+++ projects/tinymce/src/lib/template-editor/template-editor.component.ts
@@ -1,26 +1,9 @@
-import {AfterViewInit, ChangeDetectionStrategy, Component, Inject, OnInit} from '@angular/core';
+import {ChangeDetectionStrategy, Component, Inject, OnInit, ViewChild} from '@angular/core';
 import {FormControl} from '@angular/forms';
-import {COMPONENT_DATA, FieldComponent, FieldData, StorageService} from '@jaspero/form-builder';
+import {COMPONENT_DATA, FieldComponent, FieldData} from '@jaspero/form-builder';
+import {TemplateEditorSegment} from './interfaces/template-editor-segment.interface';
 import {TemplateEditorTemplate} from './interfaces/template-editor-template.interface';
-
-declare const tinymce: any;
-
-interface Segment {
-  id: string;
-  name: string;
-  content: string;
-  editors: any[];
-  group: string;
-}
-
-interface Template {
-  id: string;
-  name: string;
-  style?: string;
-  layout?: string;
-  segments: Segment[];
-  defaultSegments: string[];
-}
+import {TemplateEditorInnerComponent} from './template-editor-inner/template-editor-inner.component';
 
 interface TemplateEditorData extends FieldData {
   templates: TemplateEditorTemplate[];
@@ -34,26 +17,26 @@ interface TemplateEditorData extends FieldData {
   styleUrls: ['./template-editor.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class TemplateEditorComponent extends FieldComponent<TemplateEditorData> implements OnInit, AfterViewInit {
+export class TemplateEditorComponent extends FieldComponent<TemplateEditorData> implements OnInit {
   constructor(
-    @Inject(COMPONENT_DATA) public cData: TemplateEditorData,
-    private storage: StorageService
+    @Inject(COMPONENT_DATA) public cData: TemplateEditorData
   ) {
     super(cData);
   }
 
-  templates: Template[];
+  @ViewChild(TemplateEditorInnerComponent)
+  innerComponent: TemplateEditorInnerComponent;
+
+  templates: TemplateEditorTemplate[];
   templateControl: FormControl;
-  template: Template;
-  segments: Segment[];
-  value: string;
-  mainEl: HTMLDivElement;
+  template: TemplateEditorTemplate;
+  segments: TemplateEditorSegment[];
   segmentSelect: {
     grouped: Array<{
       group: string;
-      segments: Segment[];
+      segments: TemplateEditorSegment[];
     }>;
-    root: Segment[];
+    root: TemplateEditorSegment[];
   };
 
 
@@ -67,10 +50,10 @@ export class TemplateEditorComponent extends FieldComponent<TemplateEditorData> 
     } = this.cData.control.value;
 
     if (template) {
-      this.template = this.templates.find(it => it.id === template) as Template;
+      this.template = this.templates.find(it => it.id === template) as TemplateEditorTemplate;
       this.templateControl = new FormControl(template);
     } else {
-      this.template = this.templates.find(it => it.id === this.cData.defaultTemplate) as Template;
+      this.template = this.templates.find(it => it.id === this.cData.defaultTemplate) as TemplateEditorTemplate;
       this.templateControl = new FormControl(this.template.id);
     }
 
@@ -80,7 +63,6 @@ export class TemplateEditorComponent extends FieldComponent<TemplateEditorData> 
         .map(segment => this.template.segments.find(it => it.id === segment))
     )
       .map(data => ({...data}));
-    this.value = this.template.layout || `<div class="main-content"></div>`;
     this.segmentSelect = this.template.segments.reduce((acc: any, cur: any) => {
 
       if (!cur.group) {
@@ -110,95 +92,11 @@ export class TemplateEditorComponent extends FieldComponent<TemplateEditorData> 
     }
   }
 
-  ngAfterViewInit() {
-    this.mainEl = document.querySelector('.main-content') as HTMLDivElement;
-    this.compileSegments();
-  }
-
-  addSegment(segment: Segment) {
-
+  addSegment(segment: TemplateEditorSegment) {
     const seg = {...segment};
-
     this.segments.push(seg);
-    this.compileSegment(seg);
+    this.innerComponent.compileSegment(seg);
     this.updateData();
-  }
-
-  compileSegments() {
-    this.segments.forEach(segment => {
-      this.compileSegment(segment);
-    });
-  }
-
-  compileSegment(segment: Segment) {
-    const elWrapper = document.createElement('div');
-    const el = document.createElement('div');
-    const cleanUpTrigger = document.createElement('i');
-
-    elWrapper.style.marginBottom = '1rem';
-    elWrapper.style.position = 'relative';
-
-    el.innerHTML = segment.content;
-
-    cleanUpTrigger.classList.add('material-icons');
-
-    cleanUpTrigger.style.position = 'absolute';
-    cleanUpTrigger.style.right = '-50px';
-    cleanUpTrigger.style.top = '0';
-    cleanUpTrigger.style.cursor = 'pointer';
-
-    cleanUpTrigger.innerText = 'close';
-
-    cleanUpTrigger.onclick = () => {
-      const index = this.segments.findIndex(it => it === segment);
-
-      this.segments.splice(index, 1);
-      this.mainEl.removeChild(cleanUpTrigger.parentElement as HTMLDivElement);
-    };
-
-    elWrapper.appendChild(el);
-    elWrapper.appendChild(cleanUpTrigger);
-
-    this.mainEl.appendChild(elWrapper);
-
-    tinymce.init({
-      target: el,
-      plugins: ['link', 'lists', 'image', 'code'],
-      toolbar: ['styleselect', 'bold italic underline', 'link image emoticons', 'align bullist numlist', 'code removeformat'].join(' | '),
-      menubar: false,
-      inline: true,
-      target_list: false,
-      object_resizing: false,
-      paste_as_text: true,
-      image_dimensions: false,
-      ...this.template.style && {content_style: this.template.style},
-      // tslint:disable-next-line:max-line-length
-      extended_valid_elements: 'svg[*],defs[*],pattern[*],desc[*],metadata[*],g[*],mask[*],path[*],line[*],marker[*],rect[*],circle[*],ellipse[*],polygon[*],polyline[*],linearGradient[*],radialGradient[*],stop[*],image[*],view[*],text[*],textPath[*],title[*],tspan[*],glyph[*],symbol[*],switch[*],use[*]',
-      images_upload_handler: (blobInfo, success, failure) => {
-
-        /**
-         * TODO:
-         * We need to give a module and document here
-         * so that images are properly removed when
-         * a document is remove
-         */
-        this.storage
-          .upload(blobInfo.filename(), blobInfo.blob())
-          .then(data => data.ref.getDownloadURL())
-          .then(url => success(url))
-          .catch(error => failure(error.toString()));
-      },
-      setup: editor => {
-        editor.on('keyup change', () => {
-          segment.content = editor.getContent();
-          this.updateData();
-        });
-      },
-      style_formats: [
-        {title: 'Image', selector: 'img', styles: {width : '100%', height: 'auto'}}
-      ],
-      ...this.cData.wysiwygConfig || {}
-    });
   }
 
   updateData() {