+++ projects/tinymce/src/lib/template-editor/template-editor.module.ts
@@ -4,10 +4,11 @@ import {MatButtonModule} from '@angular/material/button';
 import {MatMenuModule} from '@angular/material/menu';
 import {FormBuilderContextService, FormBuilderModule} from '@jaspero/form-builder';
 import {TranslocoModule} from '@ngneat/transloco';
+import {TemplateEditorInnerComponent} from './template-editor-inner/template-editor-inner.component';
 import {TemplateEditorComponent} from './template-editor.component';
 
 @NgModule({
-  declarations: [TemplateEditorComponent],
+  declarations: [TemplateEditorComponent, TemplateEditorInnerComponent],
   imports: [
     CommonModule,