+++ projects/page-builder/CHANGELOG.md
@@ -1,3 +1,70 @@
+## @jaspero/fb-page-builder [2.15.26](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-page-builder@2.15.25...@jaspero/fb-page-builder@2.15.26) (2021-04-13)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.23.0
+
+## @jaspero/fb-page-builder [2.15.25](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-page-builder@2.15.24...@jaspero/fb-page-builder@2.15.25) (2021-04-13)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.22.5
+
+## @jaspero/fb-page-builder [2.15.24](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-page-builder@2.15.23...@jaspero/fb-page-builder@2.15.24) (2021-04-13)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.22.4
+
+## @jaspero/fb-page-builder [2.15.23](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-page-builder@2.15.22...@jaspero/fb-page-builder@2.15.23) (2021-04-13)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.22.3
+
+## @jaspero/fb-page-builder [2.15.22](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-page-builder@2.15.21...@jaspero/fb-page-builder@2.15.22) (2021-04-13)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.22.2
+
+## @jaspero/fb-page-builder [2.15.21](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-page-builder@2.15.20...@jaspero/fb-page-builder@2.15.21) (2021-04-09)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.22.1
+
+## @jaspero/fb-page-builder [2.15.20](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-page-builder@2.15.19...@jaspero/fb-page-builder@2.15.20) (2021-03-16)
+
+
+### Bug Fixes
+
+* **blocks:** reduced z-index ([29e3c37](https://github.com/Jaspero/schema-forms/commit/29e3c378b0886db4a3a20c35ad4c6179bc994240))
+
 ## @jaspero/fb-page-builder [2.15.19](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-page-builder@2.15.18...@jaspero/fb-page-builder@2.15.19) (2021-03-06)